Add config option to for deleting vendor directories

+++ src/Mover.php
@@ -90,7 +90,9 @@ class Mover
             $this->movedPackages[] = $package->config->name;
         }
 
-        $this->deletePackageVendorDirectories();
+        if (!isset($this->config->delete_vendor_directories) || $this->config->delete_vendor_directories === true) {
+	        $this->deletePackageVendorDirectories();
+        }
     }
 
     /**
@@ -137,6 +139,9 @@ class Mover
     {
         foreach ($this->movedPackages as $movedPackage) {
             $packageDir = '/vendor/' . $movedPackage;
+            if (is_link($packageDir)) {
+            	continue;
+            }
             $this->filesystem->deleteDir($packageDir);
         }
     }
